Mapping endpoint POST /client to create a valid client.
Removing BrazilFullName annotation temporarily until it resolves issue with handleMethodArgumentNotValid that is not returning custom message when used.

Change the validation annotations from NotNull to NotEmpty, also validate when empty.

Event creation and listener to add header location every time a resource is created.

Creation of the findByEmail method to fetch a client by email, it is being used to throw an exception when there is an attempt to save a client with the same email.

Changed ApiError class to support Api validation error.